Fix update workflow
There are some pretty small but big changes to the update workflow. I'm doing this because Mojang is doing breaking changes in patch versions, which is fine but it means that version numbers are less meaningful. Since the only time I actually override them is when breaking changes happen, it's easy enough to just force the values when relevant.

Also removed the Neoforge dependency from the Neoforge TOML. I'm assuming that Neoforge itself has a check so it's not totally necessary. If this bites me down the line then I'll fix it.