Fix options category localisation
My decision to add a hyphen to the mod ID rears its head. Now there is one, and only one, string key that us hyphenated. I might have to look into changing the mod ID entirely. I think the only place that requires it to be hyphenated is the config code, which could just as easily be special without affecting anything else.